Sort breakdown rows by sub-path on the dag.
Previously, the rows inherited the sort order using all parts of the dag/chain.
This didn't make sense when one of the parts isn't present in the display (because it is shown across the breakdown).

This change requires setting up explicit structs for all breakdown variations to give the full type information.
I'm assuming this can be generalized so we can use a macro to generate these structs later.