Display the aggregate as the header on 'show_aggregate'.
This is a small quality of output improvement, but requires some new internal modelling for rendering.
We add `Value::Overflow` and `Value::Skip`, as well as other fallout accounting.

The idea is an overflow value will bleed into the subsequent columns without affecting their column width in the typical fashion.
Instead, the remainder from the overflow is applied only to stored column width the last column, if applicable.
In the case the overflow is less than the combined total of the columns it occupies, then we remember the width to use when printing the overflow value itself.