Reformulate value rendering for consistency.
Previously, we tried to linearly scale between the value domain to the rendering range.
However, this had problems in various situations, such as empty or size 1 domains.
This also could be confusing, because a domain that spans negative and positive doesn't
really visualize that nuance in the rendering.

The new framing is to linearly scale from [0, max(abs(domain))].
This always sets up a consistent way to render the values.
Additionally, we use a separate marker for negatives.