Refactor datasets to avoid the awkward "value dimension".
This idea to allow dataset re-use hasn't fully worked out yet.
One problem in particular is the value dimension which sits outside the dataset.
This change refactors datasets to remove the extra value dimension as a next step towards dataset reuse.

There is more work to come to wrap this up.