Add ability to view the value behind a breakdown.
Previously, we said "breakdowns are always 'counts'".
However, there are times when we want to render the value under breakdown.
This change allows that.

As a part of this change, we rename most of the dataset -> view functions in a consistent manner.
We also drop the "without" style views, and use a slight re-design on view pre-headers.