Use features to implement different flavoured view generation.
This is a design choice to ~force users to chose a mutually exclusive feature flag.
We do so to best support the various use cases (either primitive based or smart pointer based) without
outright prescribing one or the other.

Unfortunately, this complicates flat's build process, so an explicit build.sh is added to help
ensure all of flat builds and tests correctly.